Move some of startup logic to low.PeerGroup
